+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * added support for tribe shortcodes
+* graceful failure if The Events Calendar is not active
 
 #### 4.6.2 / 2017-01-20
 * added CSS selector to override Customizer month view